Feature/interview progress (#47)
* Minor Adjusts

* some adjusts

* Finished

* First workaround in order to execute ProgressTo

* Change when set the progress

* Completed

* Solving some visual details

* A little refactor

* Implement command to animate the progress bar

+++ Droid/CustomRenderers/ColorProgressBarRenderer.cs
@@ -81,7 +81,7 @@ namespace Steamboat.Mobile.Droid.CustomRenderers
             }
             
 
-            Control.ScaleY = element.ProgressBarHeigth; //Changes the height
+            Control.ScaleY = (float)element.ProgressBarHeigth; //Changes the height
         }
     }
 }
